Reworked AntiBreach
Should be better?

The new algorithm is similar to the old one, but with nav method added in.
Anyway it's not even correct to just check 2D distance between exploiters and saferoom doors, not to mention the wrongly predicted location based on **FULL** velocity. Traces are preferred to reliably detect if player will stuck in doors when spawned.